Hardvest Finance - Update vault info when hardwork and code refactoring
