+++ subgraphs/harvest-finance/src/mappings/VaultListener.ts
@@ -5,19 +5,19 @@ import {
   StrategyAnnounced as StrategyAnnouncedEvent,
   StrategyChanged as StrategyChangedEvent,
   Transfer as TransferEvent,
-  Withdraw as WithdrawEvent
+  Withdraw as WithdrawEvent,
+  DoHardWorkCall
 } from "../../generated/ControllerListener/VaultContract"
 import { Vault as VaultContract } from "../../generated/ControllerListener/Vault";
 import { getOrCreateToken } from './../entities/Token'
 import { Vault, Token } from "../../generated/schema";
 import { WETH_ADDRESS } from './../constant'
-import { getOrCreateVault } from './../entities/Vault'
+import { getOrCreateVault, updateVaultPrices } from './../entities/Vault'
 import { getOrCreateDeposit } from './../entities/Transaction'
 import { getOrCreateToken } from './../entities/Token'
 import { depositUpdateMetrics } from './../entities/Metrics'
 import * as constants from "./../constant";
 import { getUsdPricePerToken } from "./../Prices";
-import { readValue } from "../utils/contracts";
 
 export function handleDeposit(event: DepositEvent): void {
 
@@ -53,23 +53,22 @@ export function handleDeposit(event: DepositEvent): void {
   deposit.amountUSD = depositAmountUSD;
   deposit.save();
 
-  let vault_contract = VaultContract.bind(vaultAddress);
+  
   vault.inputTokenBalance = vault.inputTokenBalance.plus(amount);
-  vault.outputTokenSupply = readValue<BigInt>(vault_contract.try_totalSupply(), constants.BIGINT_ZERO);
-
-  vault.pricePerShare = readValue<BigInt>(vault_contract.try_getPricePerFullShare(), constants.BIGINT_ZERO).toBigDecimal();
-  vault.save();
-
-  vault.totalValueLockedUSD = (<BigDecimal> vault.pricePerShare)
-    .times(inputTokenPrice.usdPrice)
-    .times((<BigInt> vault.outputTokenSupply).toBigDecimal())
-    .div(inputTokenDecimals.toBigDecimal())
-    .div(inputTokenDecimals.toBigDecimal())
-    .div(inputTokenPrice.decimals.toBigDecimal());
+  
+  updateVaultPrices(vault);
 
   vault.save();
 
   depositUpdateMetrics(event, vault);
   
+}
+
+export function handleDoHardWorkCall(call: DoHardWorkCall): void {
+  const vaultAddress = call.to;
+
+  let vault = getOrCreateVault(vaultAddress, call.block);
+  updateVaultPrices(vault);
+
 }
